Mean Time To Failure (MTTF).

Mean Time To Failure (MTTF).

Usage

mttf(failures, totalTime)

Arguments

failures

Failures. A numeric value representing the number of failures for a given period.

totalTime

Total Time. A numeric value representing the total time for a given period or a numeric vector of time periods

Value

The function returns a numeric value representing the MTTF for a given period.

Examples

fail <- 5
total <- 1000
mttf(fail, total)
#> [1] 200